+++ africanus/dft/kernels.py
@@ -25,32 +25,31 @@ def im_to_vis(image, uvw, lm, frequency, dtype=None):
         out_dtype = dtype.dtype
 
     def _im_to_vis_impl(image, uvw, lm, frequency, dtype=None):
-        vis_of_im = np.zeros((uvw.shape[0], frequency.shape[0]),
-                             dtype=out_dtype)
+        nrows = uvw.shape[0]
+        nsrc = lm.shape[0]
+        nchan = frequency.shape[0]
+        ncorr = image.shape[-1]
+        vis_of_im = np.zeros((nrows, nchan, ncorr), dtype=out_dtype)
 
         # For each uvw coordinate
-        for row in range(uvw.shape[0]):
-            u, v, w = uvw[row]
+        for r in range(nrows):
+            u, v, w = uvw[r]
 
             # For each source
-            for source in range(lm.shape[0]):
-                l, m = lm[source]
+            for s in range(nsrc):
+                l, m = lm[s]
                 n = np.sqrt(1.0 - l**2 - m**2) - 1.0
 
                 # e^(-2*pi*(l*u + m*v + n*w)/c)
                 real_phase = minus_two_pi_over_c * (l * u + m * v + n * w)
 
                 # Multiple in frequency for each channel
-                for chan in range(frequency.shape[0]):
-                    p = real_phase * frequency[chan] * 1.0j
+                for nu in range(nchan):
+                    p = real_phase * frequency[nu] * 1.0j
 
-                    # Our phase input is purely imaginary
-                    # so we can can elide a call to exp
-                    # and just compute the cos and sin
-                    # @simon does this really make a difference?
-                    # I thought a complex exponential is evaluated
-                    # as a sum of sin and cos anyway
-                    vis_of_im[row, chan] += np.exp(p)*image[source, chan]
+                    for c in range(ncorr):
+                        if image[s, nu, c]:
+                            vis_of_im[r, nu, c] += np.exp(p)*image[s, nu, c]
 
         return vis_of_im
 
@@ -58,7 +57,7 @@ def im_to_vis(image, uvw, lm, frequency, dtype=None):
 
 
 @generated_jit(nopython=True, nogil=True, cache=True)
-def vis_to_im(vis, uvw, lm, frequency, dtype=None):
+def vis_to_im(vis, uvw, lm, frequency, flags, dtype=None):
     # Infer output dtype if none provided
     if is_numba_type_none(dtype):
         # Support both real and complex visibilities...
@@ -76,30 +75,42 @@ def vis_to_im(vis, uvw, lm, frequency, dtype=None):
 
         out_dtype = dtype.dtype
 
-    def _vis_to_im_impl(vis, uvw, lm, frequency, dtype=None):
-        im_of_vis = np.zeros((lm.shape[0], frequency.shape[0]),
-                             dtype=out_dtype)
+    assert np.shape(vis) == np.shape(flags)
+
+    def _vis_to_im_impl(vis, uvw, lm, frequency, flags, dtype=None):
+        nrows = uvw.shape[0]
+        nsrc = lm.shape[0]
+        nchan = frequency.shape[0]
+        ncorr = vis.shape[-1]
+
+        im_of_vis = np.zeros((nsrc, nchan, ncorr), dtype=out_dtype)
 
         # For each source
-        for source in range(lm.shape[0]):
-            l, m = lm[source]
+        for s in range(nsrc):
+            l, m = lm[s]
             n = np.sqrt(1.0 - l ** 2 - m ** 2) - 1.0
             # For each uvw coordinate
-            for row in range(uvw.shape[0]):
-                u, v, w = uvw[row]
+            for r in range(nrows):
+                u, v, w = uvw[r]
 
                 # e^(-2*pi*(l*u + m*v + n*w)/c)
                 real_phase = -minus_two_pi_over_c * (l * u + m * v + n * w)
 
                 # Multiple in frequency for each channel
-                for chan in range(frequency.shape[0]):
-                    p = real_phase * frequency[chan]
+                for nu in range(nchan):
+                    p = real_phase * frequency[nu]
 
-                    im_of_vis[source,
-                              chan] += (np.cos(p) * vis[row, chan].real -
-                                        np.sin(p) * vis[row, chan].imag)
-                    # Note for the adjoint we don't need the imaginary part
-                    # and we can elide the call to exp
+                    # do not compute if any of the correlations
+                    # are flagged (complicates uncertainties)
+                    if np.any(flags[r, nu]):
+                        continue
+
+                    for c in range(ncorr):
+                        # elide the call to exp since result is real
+                        im_of_vis[s, nu, c] += (np.cos(p) *
+                                                vis[r, nu, c].real -
+                                                np.sin(p) *
+                                                vis[r, nu, c].imag)
 
         return im_of_vis
 
@@ -111,8 +122,8 @@ _DFT_DOCSTRING = namedtuple(
 
 im_to_vis_docs = _DFT_DOCSTRING(
     preamble="""
-    Computes the discrete image to visibility mapping of an ideal
-    unpolarised interferometer :
+    Computes the discrete image to visibility mapping
+    of an ideal interferometer:
 
     .. math::
 
@@ -125,14 +136,15 @@ im_to_vis_docs = _DFT_DOCSTRING(
     ----------
 
     image : :class:`numpy.ndarray`
-        image of shape :code:`(source, chan)`
-        The Stokes I intensity in each pixel (flatten 2D array per channel).
+        image of shape :code:`(source, chan, corr)`
+        The brighness matrix in each pixel (flatten 2D array
+        per channel and corr). Note not Stokes terms
     uvw : :class:`numpy.ndarray`
-        UVW coordinates of shape :code:`(row, 3)` with
-        U, V and W components in the last dimension.
+        uvw coordinates of shape :code:`(row, 3)` with
+        u, v and w components in the last dimension.
     lm : :class:`numpy.ndarray`
-        LM coordinates of shape :code:`(source, 2)` with
-        L and M components in the last dimension.
+        lm coordinates of shape :code:`(source, 2)` with
+        l and m components in the last dimension.
     frequency : :class:`numpy.ndarray`
         frequencies of shape :code:`(chan,)`
     dtype : np.dtype, optional
@@ -145,7 +157,7 @@ im_to_vis_docs = _DFT_DOCSTRING(
     Returns
     -------
     visibilties : :class:`numpy.ndarray`
-        complex of shape :code:`(row, chan)`
+        complex of shape :code:`(row, chan, corr)`
     """
 )
 
@@ -154,8 +166,8 @@ im_to_vis.__doc__ = doc_tuple_to_str(im_to_vis_docs)
 
 vis_to_im_docs = _DFT_DOCSTRING(
     preamble="""
-    Computes visibility to image mapping of an ideal
-    unpolarised interferometer:
+    Computes visibility to image mapping
+    of an ideal interferometer:
 
     .. math::
 
@@ -168,16 +180,23 @@ vis_to_im_docs = _DFT_DOCSTRING(
     ----------
 
     vis : :class:`numpy.ndarray`
-        visibilities of shape :code:`(row, chan)`
-        The Stokes I visibilities of which to compute a dirty image
+        visibilities of shape :code:`(row, chan, corr)`
+        Visibilities corresponding to brightness terms.
+        Note the dirty images produced do not necessarily
+        correspond to Stokes terms and need to be converted.
     uvw : :class:`numpy.ndarray`
-        UVW coordinates of shape :code:`(row, 3)` with
-        U, V and W components in the last dimension.
+        uvw coordinates of shape :code:`(row, 3)` with
+        u, v and w components in the last dimension.
     lm : :class:`numpy.ndarray`
-        LM coordinates of shape :code:`(source, 2)` with
-        L and M components in the last dimension.
+        lm coordinates of shape :code:`(source, 2)` with
+        l and m components in the last dimension.
     frequency : :class:`numpy.ndarray`
         frequencies of shape :code:`(chan,)`
+    flags : :class:`numpy.ndarray`
+        Boolean array of shape :code:`(row, chan, corr)`
+        Note that if one correlation is flagged we discard
+        all of them otherwise we end up irretrievably
+        mixing Stokes terms.
     dtype : np.dtype, optional
         Datatype of result. Should be either np.float32 or np.float64.
         If ``None``, :func:`numpy.result_type` is used to infer the data type
@@ -188,7 +207,7 @@ vis_to_im_docs = _DFT_DOCSTRING(
     Returns
     -------
     image : :class:`numpy.ndarray`
-        float of shape :code:`(source, chan)`
+        float of shape :code:`(source, chan, corr)`
     """
 )